+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog for the Hermod Tracing Project
 
+## June 2026
+
+* Import `hermod-recon-framework` (was: `cardano-recon-framework`).
+
 ## March 2026
 
 * Release `trace-dispatcher-2.12.0`.